+++ flutter_reach_five_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.0.1
+
+- **FEAT**: Add `accountBlockedAfterMultipleLoginAttempts` error code to `ErrorCodesInterface`
+
 # 2.0.0
 
 - **BREAKING**: Update Flutter to >=3.41.0